requirements.txt
pdf2image==1.17.0
google-generativeai==0.5.4
PyPDF2==3.0.1
tweepy==4.14.0
